src/VerdeBordo.Application/Features/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
using MediatR;
using VerdeBordo.Core.Repositories;

namespace VerdeBordo.Application.Features.Users.Commands.DeleteUser;

public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, Unit>
{
    private readonly IUserRepository _userRepository;

    public DeleteUserCommandHandler(IUserRepository userRepository)
    {
        _userRepository = userRepository;
    }

    public async Task<Unit> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
    {
        var user = await _userRepository.GetByIdAsync(request.Id);

        if (user is null)
            throw new Exception("User not found");

        user.Delete();

        await _userRepository.UpdateAsync(user);

        return Unit.Value;
    }
}